大部分已知bug修复 (#61)
* create webproject

* create android

* 创建数据表

* test

* 数据构建完成，用户信息的增删改查

* 添加部门申请状态

* 成员列表完善

* merge

* 部门活动数据修改

* 子部门和活动中心界面与数据交互

* 头像添加接口完成

* 活动详情页面编写和数据交互

* 登录添加部门逻辑

* alpha远程服务器版本

* 通知中心，bug修复

* 通知中心完善，bug修复

+++ xmatch_web/src/main/java/com/zsq/service/impl/DepartmentServiceImpl.java
@@ -52,6 +52,7 @@ public class DepartmentServiceImpl implements DepartmentService {
             return LsyResultCode.Companion.getDEP_NOT_EXIT();
         }
         resDep.setDepSummary(dep.getDepSummary());
+        resDep.setEmergencyPhone(dep.getEmergencyPhone());
         return LsyResultCode.Companion.getSUCCESS();
     }
 
@@ -125,6 +126,14 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     public int deleteChildDepartment(long childDepartmentId) {
+        ChildDepartment childDepartment = childDepartmentRepository.findOne(childDepartmentId);
+        if (childDepartment == null) {
+            return WyyResultCode.Companion.getCHILD_DEP_NOT_EXIST();
+        }
+        Department department = repository.findOne(childDepartment.getDepartmentId());
+        if(department!=null) {
+            department.setChildDepNum(department.getChildDepNum()-1);
+        }
         childDepartmentRepository.delete(childDepartmentId);
         return WyyResultCode.Companion.getSUCCESS();
     }
